french translation, testing, & further redesign
- French translation courtesy of Jumpman added to the patcher
- New testing feature for translators which opens all the program's popup windows one-by-one
- Cleaned up patcher design; confirmation dialogs now appear if you attempt patching a ROM with a sample rate other than default
- Fix incorrect icons on certain dialogs
- Translate title for ROM open dialog
- Fix error messages that weren't being localized
- Add French translation credit

+++ The MOTHER 3 HQ Audio Patcher/src/main/java/org/musictheorist/mother3hqaudiopatcher/ui/popups/TextBox.java
@@ -28,9 +28,9 @@ public final class TextBox {
         return text;
     }
 
-    private static double getSizeFromKey(String key, ResourceBundle lang) {
+    private static int getSizeFromKey(String key, ResourceBundle lang) {
         String value = lang.getString(key);
-        return Double.parseDouble(value);
+        return Integer.parseInt(value);
     }
 
     private static void setSizes(TextArea textBox, double width, double height) {
@@ -55,8 +55,8 @@ public final class TextBox {
         String contents = lang.getString(contentsKey);
         TextArea textBox = createTextBox(contents, true);
 
-        double width = getSizeFromKey(widthKey, lang);
-        double height = getSizeFromKey(heightKey, lang);
+        int width = getSizeFromKey(widthKey, lang);
+        int height = getSizeFromKey(heightKey, lang);
         textBox.setPrefSize(width, height);
 
         return textBox;
